refactor: Ordnerstruktur auf core/shared/features umstellen
core/    Domaenen-Modelle, App-weite Services und der HTTP-Interceptor
shared/  wiederverwendbare UI-Bausteine (modal, player-item, theme-toggle),
         die EuroPipe und die FormattedNumber-Direktive
features/ die fachlichen Ansichten help, login und market-overview

Reine Verschiebung samt nachgezogener Import-Pfade; git erkennt alle 47
Dateien als Rename. no-decimals.pipe.ts heisst jetzt euro.pipe.ts, passend
zur enthaltenen EuroPipe. Kein Verhalten geaendert.
