+++ src/app/core/models/kickbase-gift.ts
