+++ src/app/core/models/kickbase-group.ts
