+++ src/app/core/models/kickbase-league.ts
