+++ src/app/core/models/kickbase-market.ts
