+++ src/app/core/models/kickbase-player-stats.ts
