+++ src/app/core/models/kickbase-player.ts
