+++ src/app/core/services/api.service.spec.ts
@@ -2,7 +2,7 @@ import { TestBed } from '@angular/core/testing';
 import { HttpTestingController, provideHttpClientTesting } from '@angular/common/http/testing';
 import { provideHttpClient } from '@angular/common/http';
 import { ApiService } from './api.service';
-import { KickbaseLeague } from '../model/kickbase-league';
+import { KickbaseLeague } from '../models/kickbase-league';
 
 describe('ApiService', () => {
   let service: ApiService;
