+++ src/app/core/services/theme.service.spec.ts
