+++ src/app/features/market-overview/market-overview.component.scss
