+++ src/app/shared/components/player-item/player-item.component.scss
