+++ src/app/shared/components/theme-toggle/theme-toggle.component.ts
@@ -1,6 +1,6 @@
 import { ChangeDetectionStrategy, Component, computed, inject } from '@angular/core';
 import { AngularSvgIconModule } from 'angular-svg-icon';
-import { THEME_MODE_ORDER, ThemeMode, ThemeService } from '../../services/theme.service';
+import { THEME_MODE_ORDER, ThemeMode, ThemeService } from '../../../core/services/theme.service';
 
 const THEME_ICONS: Record<ThemeMode, string> = {
   system: 'bi bi-circle-half',
