+++ src/app/shared/directives/formatted-number.directive.ts
