+++ src/app/shared/pipes/euro.pipe.ts
